// src/app/features/hist-real-pnl-query/models/hist-real-pnl-query-request.model.ts
export interface HistRealPnlQueryRequest {
  APISERVER: string;
  sid: string;
  sip: string;
  Invscode: string;
  comp: string;
  bhno: string;
  cseq: string;
  stock: string;
  stockName: string;
  ttype: string;
  ctype: string;
  bdate: string;
  edate: string;
  action: string;
  action1: string;
  action4: string;
}